- config buttons reset to default não está funcionando - na configurações de botões, não pode mudar BACK, que é usado para sair para outra tecla - não deixar sair da configuração de botões de pulo, tiro ou START estão como -1 - crash na luta contra seahorse - melhorar AI Seahorsebot (virar para player para atacar) - peixe azul da fase seahorse está piscando

+++ inputlib.h
@@ -27,7 +27,7 @@ public:
 
     int get_joysticks_number();
 
-    std::string get_key_name(Uint8 key);
+    std::string get_key_name(int key);
